requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
cmake==3.25.0
contourpy==1.1.0
cycler==0.11.0
filelock==3.9.0
fonttools==4.40.0
fsspec==2023.6.0
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.1
imgaug==0.4.0
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.3.1
kiwisolver==1.4.4
kornia==0.7.0
lazy_loader==0.3
lit==15.0.7
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.2.1
networkx==3.0
nibabel==5.1.0
numpy==1.23.1
opencv-python==4.8.0.74
packaging==23.1
Pillow==10.0.0
pyparsing==3.1.0
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.28.1
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
shapely==2.0.1
six==1.16.0
sympy==1.11.1
threadpoolctl==3.1.0
tifffile==2023.7.4
timm==0.9.5
# torch==2.0.1+cu118
torchsummary==1.5.1
# torchvision==0.15.2+cu118
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.7.1
urllib3==1.26.13
zipp==3.15.0
MedPy==0.4.0
